PowerBISummarizer/cloud-api/app/schemas.py
from datetime import datetime
from pydantic import BaseModel, EmailStr


class LoginRequest(BaseModel):
    email: EmailStr
    password: str


class TokenResponse(BaseModel):
    access_token: str
    token_type: str = "bearer"
    expires_in: int


class UserOut(BaseModel):
    id: int
    email: EmailStr
    role: str
    created_at: datetime

    class Config:
        orm_mode = True


class LayerOut(BaseModel):
    id: int
    name: str
    schema: str
    srid: int
    geom_type: str

    class Config:
        orm_mode = True
