PowerBISummarizer/cloud-api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.34
psycopg[binary]==3.1.12
python-jose==3.3.0
passlib[bcrypt]==1.7.4
